requirements.txt
fastapi==0.101.0
gunicorn==21.2.0
uvicorn==0.23.2
asyncpg==0.28.0
redis==4.6.0
fastapi-cache2[redis]==0.2.1
orjson==3.9.7
pydantic-settings==2.0.3